Move _hide_source_ranges() to __init__.py
